Clean & fix conversation settings crash

+++ SignalServiceKit/src/Storage/AxolotlStore/OWSPrimaryStorage+PreKeyStore.h
@@ -10,7 +10,7 @@ NS_ASSUME_NONNULL_BEGIN
 @interface OWSPrimaryStorage (PreKeyStore) <PreKeyStore>
 
 - (NSArray<PreKeyRecord *> *)generatePreKeyRecords;
-- (NSArray<PreKeyRecord *> *)generatePreKeyRecords:(NSUInteger)batchSize;
+- (NSArray<PreKeyRecord *> *)generatePreKeyRecords:(int)batchSize;
 - (void)storePreKeyRecords:(NSArray<PreKeyRecord *> *)preKeyRecords NS_SWIFT_NAME(storePreKeyRecords(_:));
 
 @end
